requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.10.1
sqlparse==0.5.3
